fix(auth): move the WWW-Authenticate parser out from behind the reqwest gate  [TR-429]
Second instance of the TR-428 gap, found by applying the same completeness
test to the other three signers.

`DigestBuildParams` is filled ENTIRELY from a parsed `WWW-Authenticate`
challenge — realm, nonce, qop, algorithm and opaque all come from it. The
parser sat in `signers.rs`, which is `#[cfg(feature = "reqwest")]`, so a
browser caller could reach `digest_build_authorization` but had no way to
produce a single one of its inputs without re-implementing RFC 7235 §4.1
parsing in TypeScript.

Moved, ungated: `parse_challenges`, `find_digest_challenge`,
`split_challenge_parts`, `parse_challenge_part`.

Two of the rules those implement are ones tropel has already shipped wrong
once, per their own comments: a challenge listing Digest AFTER Basic in one
header value was silently skipped (backlog line 176), and a naive
`split(',')` truncated a quoted qop list so a server advertising
`qop="auth, auth-int"` lost auth-int. A TypeScript re-derivation gets to
rediscover both.

The three pure parser tests move with the code, so they now run in the
browser configuration rather than only with the reqwest feature — the
precise reason the TR-428 gap went unnoticed. `cargo test -p tropel-auth
--no-default-features` goes from 49 to 53. The new
`digest_challenge_is_fully_parseable_without_the_reqwest_feature` walks
challenge -> params -> Authorization using nothing but `builders`, and is
enforced by compilation rather than assertion.

No behaviour change: this is a move plus a visibility change.

Note, NOT fixed here: `split_challenge_parts` toggles its quote flag on
every `"` with no regard for backslash escapes, so a realm containing `\"`
corrupts the realm AND swallows every parameter after it (the nonce is
simply lost). The write side already escapes correctly, so tropel escapes
on output and does not unescape on input. That is a behaviour change with
its own tests and is deliberately left to TR-430 rather than mixed into a
pure move.
